refactor: modularize symbolic ODE generation (#103)
