+++ src/cubie/systemmodels/__init__.py
@@ -5,5 +5,5 @@ ordinary differential equation systems on CUDA devices.
 """
 
 from cubie.systemmodels.systems import ThreeChamberModel, GenericODE, Decays
-from cubie.systemmodels.symbolic import SymbolicODESystem
-__all__ = ["ThreeChamberModel", "GenericODE", "Decays", "SymbolicODESystem"]
+from cubie.systemmodels.symbolic import SymbolicODE
+__all__ = ["ThreeChamberModel", "GenericODE", "Decays", "SymbolicODE"]
