+++ src/cubie/systemmodels/symbolic/__init__.py
@@ -1,9 +1,10 @@
 """Symbolic system building utilities."""
 
-from cubie.systemmodels.symbolic.symbolic import (
+from cubie.systemmodels.symbolic.symbolicODE import (
+    SymbolicODE,
     SymbolicODESystem,
-    setup_system,
 )
+from cubie.systemmodels.symbolic.parser import setup_system
 from cubie.systemmodels.symbolic.math_functions import (
     exp_,
     sin_,
@@ -14,6 +15,7 @@ from cubie.systemmodels.symbolic.math_functions import (
 )
 
 __all__ = [
+    "SymbolicODE",
     "SymbolicODESystem",
     "setup_system",
     "exp_",
